requirements.txt
torch
transformers
diffusers
accelerate
bitsandbytes
sentencepiece
Pillow
tk
